+++ node/src/components/plots/Metric.tsx
