+++ node/src/components/setup/ValueConfig.tsx
@@ -1,15 +1,16 @@
 import { Box, Typography, Chip, useTheme } from "@mui/material";
-import { InputBlock } from "./InputBlock";
+import { InputBlock } from "../utils/InputBlock";
 
-interface VariableConfigProps {
-  inputVar: SamplingInputsState;
+const VariableConfig = ({
+  inputVar,
+  index,
+  handleInputChange,
+}: {
+  inputVar: SingleJobConfig;
   index: number;
-  handleInputChange: (index: number, field: fieldType, value: string) => void;
-}
-
-const VariableConfig = (props: VariableConfigProps) => {
+  handleInputChange: (index: number, field: string, value: string) => void;
+}) => {
   const theme = useTheme();
-  const { inputVar, index, handleInputChange } = props;
 
   return (
     <Box
@@ -38,17 +39,13 @@ const VariableConfig = (props: VariableConfigProps) => {
             borderRadius: "8px",
             backgroundColor: theme.palette.primary.main,
           }}
-        ></Chip>
+        />
       </Typography>
       <InputBlock
-        name="Start"
-        value={inputVar.start}
-        onChange={(n) => handleInputChange(index, "start", n as string)}
-      />
-      <InputBlock
-        name="End"
-        value={inputVar.end}
-        onChange={(n) => handleInputChange(index, "end", n as string)}
+        name="Value"
+        type="number"
+        value={inputVar.value !== undefined ? inputVar.value : NaN}
+        onChange={(n) => handleInputChange(index, "value", n as string)}
       />
     </Box>
   );
