+++ node/src/views/OutputSetup.tsx
@@ -9,8 +9,8 @@ import {
   useTheme,
 } from "@mui/material";
 import { InfoOutline } from "@mui/icons-material";
-import CustomTooltip from "./CustomTooltip";
-import SelectQoIDocument from "./documents/SelectQoIDocument";
+import CustomTooltip from "../components/utils/CustomTooltip";
+import SelectQoIDocument from "../components/documents/SelectQoIDocument";
 import { useMMUXContext } from "../context/MMUXContext";
 
 interface UQSetupProps {
